+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/rules/IndentationRule.kt
@@ -12,6 +12,7 @@ import com.pinterest.ktlint.rule.engine.core.api.ElementType.BLOCK
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.BLOCK_COMMENT
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.BODY
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.CALL_EXPRESSION
+import com.pinterest.ktlint.rule.engine.core.api.ElementType.CATCH
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.CLASS
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.CLASS_BODY
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.CLOSING_QUOTE
@@ -26,6 +27,7 @@ import com.pinterest.ktlint.rule.engine.core.api.ElementType.DOT_QUALIFIED_EXPRE
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.ELSE
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.ELVIS
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.EQ
+import com.pinterest.ktlint.rule.engine.core.api.ElementType.FINALLY
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.FOR
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.FUN
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.FUNCTION_LITERAL
@@ -61,6 +63,8 @@ import com.pinterest.ktlint.rule.engine.core.api.ElementType.SUPER_TYPE_CALL_ENT
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.SUPER_TYPE_ENTRY
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.SUPER_TYPE_LIST
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.THEN
+import com.pinterest.ktlint.rule.engine.core.api.ElementType.TRY
+import com.pinterest.ktlint.rule.engine.core.api.ElementType.TYPEALIAS
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.TYPE_ARGUMENT_LIST
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.TYPE_CONSTRAINT
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.TYPE_CONSTRAINT_LIST
@@ -109,8 +113,7 @@ import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
 import org.jetbrains.kotlin.psi.KtStringTemplateExpression
 import org.jetbrains.kotlin.psi.psiUtil.leaves
 import org.jetbrains.kotlin.psi.psiUtil.parents
-import java.util.Deque
-import java.util.LinkedList
+import java.util.*
 
 private val LOGGER = KotlinLogging.logger {}.initKtLintKLogger()
 
@@ -296,7 +299,8 @@ public class IndentationRule :
             node.elementType == KDOC ->
                 visitKdoc(node)
 
-            node.elementType == PROPERTY_ACCESSOR ->
+            node.elementType == PROPERTY_ACCESSOR ||
+                node.elementType == TYPEALIAS ->
                 visitPropertyAccessor(node)
 
             node.elementType == FOR ||
@@ -309,7 +313,11 @@ public class IndentationRule :
             node.elementType == NULLABLE_TYPE ->
                 visitNullableType(node)
 
-            node.elementType == DESTRUCTURING_DECLARATION -> visitDestructuringDeclaration(node)
+            node.elementType == DESTRUCTURING_DECLARATION ->
+                visitDestructuringDeclaration(node)
+
+            node.elementType == TRY ->
+                visitTryCatchFinally(node)
 
             else -> {
                 LOGGER.trace { "No processing for ${node.elementType}: ${node.textWithEscapedTabAndNewline()}" }
@@ -763,6 +771,36 @@ public class IndentationRule :
                     lastChildIndent = "",
                 )
             }
+    }
+
+    private fun visitTryCatchFinally(node: ASTNode) {
+        // Inner indent contexts in reversed order
+        var nextToAstNode = node.lastChildLeafOrSelf()
+        node
+            .findChildByType(FINALLY)
+            ?.let { finally ->
+                nextToAstNode = startIndentContext(
+                    fromAstNode = finally,
+                    toAstNode = nextToAstNode,
+                    childIndent = "",
+                ).prevCodeLeaf()
+            }
+
+        node
+            .findChildByType(CATCH)
+            ?.let { catch ->
+                nextToAstNode = startIndentContext(
+                    fromAstNode = catch,
+                    toAstNode = nextToAstNode,
+                    childIndent = "",
+                ).prevCodeLeaf()
+            }
+
+        startIndentContext(
+            fromAstNode = node,
+            toAstNode = nextToAstNode,
+            lastChildIndent = "",
+        )
     }
 
     private fun ASTNode.skipLeadingWhitespaceCommentsAndAnnotations(): ASTNode {